Move HostDetails to page dir and fix storybook
