+++ assets/js/pages/HostDetails/HostDetails.jsx
