+++ assets/js/pages/HostDetails/HostDetailsPage.jsx
