+++ assets/js/pages/HostDetails/ProviderDetails.jsx
