+++ assets/js/pages/HostDetails/index.js
