+++ assets/js/pages/HostDetails/tableConfigs.jsx
