+++ assets/js/trento.jsx
@@ -15,9 +15,10 @@ import HostsList from '@components/HostsList';
 import ClustersList from '@pages/ClusterDetails/ClustersList';
 import ClusterDetailsPage from '@pages/ClusterDetails';
 import ClusterSettingsPage from '@components/ClusterSettingsPage';
-import { ExecutionResultsPage } from '@components/ExecutionResults';
+import { ExecutionResultsPage } from '@pages/ExecutionResults';
+import CheckResultDetailPage from '@pages/ExecutionResults/CheckResultDetail';
 import SapSystemsOverviewPage from '@components/SapSystemsOverview';
-import HostDetailsPage, { HostSettingsPage } from '@components/HostDetails';
+import HostDetailsPage, { HostSettingsPage } from '@pages/HostDetails';
 import DatabasesOverviewPage from '@pages/DatabasesOverview';
 import ChecksCatalogPage from '@pages/ChecksCatalog';
 import NotFound from '@components/NotFound';
@@ -29,7 +30,6 @@ import Login from '@components/Login';
 import { me } from '@lib/auth';
 import { networkClient } from '@lib/network';
 import Guard from '@components/Guard';
-import CheckResultDetailPage from '@components/ExecutionResults/CheckResultDetail';
 import { TARGET_CLUSTER, TARGET_HOST } from '@lib/model';
 import DatabaseDetails from './pages/DatabaseDetails';
 import SapSystemDetails from './components/SapSystemDetails/SapSystemDetails';
